NyaySahayak/requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.30.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.16.3
opencv-python==4.9.0.80

# ML & LLM (REQUIRED)
torch==2.1.2
transformers==4.37.0
sentence-transformers==2.3.1
accelerate==0.25.0

# NLP & Embeddings
langchain==0.1.4
langchain-community==0.0.16

# Vector Store
faiss-cpu==1.7.4
pgvector==0.4.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Multilingual
langdetect==1.0.9
# indicnlp-corpus==0.1.0  # Install separately for IndicTrans2

# Utilities
numpy==1.26.3
pandas==2.1.4
requests==2.31.0
python-multipart==0.0.6
tiktoken==0.5.2

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
